Agregado el CanalConsola, para enviar mensajes usando el método existente, por ejemplo en las exceptions
